+++ src/tests/finite_field_tests.rs
@@ -1,18 +1,18 @@
 #[cfg(test)]
 mod math_tests {
-    use num_bigint::BigInt;
+    use rug::Integer;
     use safhe_house::math::finite_field::{primitive_nth_root_of_unity, square_root_mod_p};
 
     #[test]
     fn tonelli_shanks() {
-        assert_eq!(square_root_mod_p(&BigInt::from(5), &BigInt::from(41)), BigInt::from(28)); 
+        assert_eq!(square_root_mod_p(&Integer::from(5), &Integer::from(41)), Integer::from(28)); 
     }
 
 
     #[test]
     fn primitive_nth_root_of_unity_test() {
-        let w = primitive_nth_root_of_unity(&BigInt::from(7681), 4);
-        assert!(w == BigInt::from(3383) || w == BigInt::from(4298) || w == BigInt::from(7680));
+        let w = primitive_nth_root_of_unity(&Integer::from(7681), 4);
+        assert!(w == Integer::from(3383) || w == Integer::from(4298) || w == Integer::from(7680));
     }
 
 }
